fix(ci): drop redundant u32 cast flagged by clippy 1.97
clippy::unnecessary_cast fires on `tier.perks.len() as u32` because
Vec::len() already returns u32 in the Soroban SDK. The lint is newer than
the toolchain this was developed against (1.94), so it only surfaced in CI,
which installs floating `stable` and runs clippy with -D warnings.

+++ contracts/shade/src/components/backer_rewards.rs
@@ -268,7 +268,7 @@ pub fn select_backer_reward_tier(env: &Env, backer: Address, campaign_id: u64, t
         backer,
         tier_index,
         tier.min_pledge,
-        tier.perks.len() as u32,
+        tier.perks.len(),
         env.ledger().timestamp(),
     );
 }
